Residuals and Pickles
Added residual connction to last LSTM layer in MDN. Modified the training procedure to use a single pickle file of data rather than multiple CSV files. This gives us the advantage of being able to train with lineStrokes data, which is data that's all written on the same line. The previous set of cleaned training data consisted of strokes that could carry over lines, which led to wonky predictions.