Model/Plotting Modifications
Took off one of the LSTM layers from the model (last layer). The LSTMCell still exists in the computation graph, but it isn't connected to anything. Modified attention weight plotting in the trainer to save images with higher DPI, better scaling, correct min/max limits, and with letters mapped to indices on the y-axis of the plot.

Need to remove the extra LSTM/recurrent state from the computation graph and modify the cyclical running to include ASCII biases.