Attention Window Changes
Changed attention window in the model to correctly break up alpha, beta, kappa values. Added phi as an output to the model (alignment of indices of text with writing timesteps). Made phi plotted. Use the same data from the test data to generate images during training (so that the network's progress on a single item can be viewed as training progresses). Modified data handler to add spaces to the end of ASCII data. Removed '@' character and zero-hot encoding from the data handler's alphabet.
Still not getting great alignment on characters during training. The ~linear streaks that we intuitively expect appear on the phi plots early in training, but these eventually converge to vertical lines.