Increased Window Cell Responsibility, Added '@'
Changed Window Cell to produce phi at the given timestep for each batch of training data. When unrolled in time, the window cell returns phi(t, u). Added '@' symbol as a zero-hot vector to the ASCII returned from the data handler. Previous experiments showed that attention was being spread to just the spaces, ' ', that were being used to pad the length of the ASCII sequences.